telegram-bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
requests==2.31.0